# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Timetable API
  version: 0.1.0
  description: |
    REST API of the university timetabling system.

    Authentication: `POST /api/auth/login` sets the `tt_access` (JWT, 15 min) and `tt_refresh`
    (30 days) HttpOnly cookies and returns the access token, which API clients may send as
    `Authorization: Bearer <token>`. Reads of reference data need any authenticated user,
    changes need the `admin` role. Every change is written to the audit log with the
    authenticated user as the actor.

    Errors always have the body `{"error": {"code": "...", "message": "..."}}`:
    400 `invalid_request` (malformed JSON, unknown fields, bad path/query parameter),
    401 `unauthorized`, 403 `forbidden`, 404 `not_found`,
    409 `already_exists` (unique key), `in_use` (deleting a referenced record), `schedule_conflict`,
    422 `validation_failed`, `invalid_reference` (referenced record missing), `constraint_violation`,
    503 `no_database`.
servers:
  - url: /
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/rooms/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [rooms]
      operationId: getRoom
      summary: Get a room
      responses:
        "200": { $ref: "#/components/responses/Room" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [rooms]
      operationId: updateRoom
      summary: Replace a room
      requestBody: { $ref: "#/components/requestBodies/RoomInput" }
      responses:
        "200": { $ref: "#/components/responses/Room" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      tags: [rooms]
      operationId: deleteRoom
      summary: Delete a room
      description: Fails with 409 `in_use` while schedule assignments use it.
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/rooms/{id}/availability:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [rooms]
      operationId: getRoomAvailability
      summary: Availability exceptions of a room
      responses:
        "200": { $ref: "#/components/responses/Availability" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [rooms]
      operationId: putRoomAvailability
      summary: Replace the availability exceptions of a room
      requestBody: { $ref: "#/components/requestBodies/Availability" }
      responses:
        "200": { $ref: "#/components/responses/Availability" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }

  # ------------------------------------------------------------------ groups
...
  /api/teachers/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [teachers]
      operationId: getTeacher
      summary: Get a teacher
      responses:
        "200": { $ref: "#/components/responses/Teacher" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [teachers]
      operationId: updateTeacher
      summary: Replace a teacher
      requestBody: { $ref: "#/components/requestBodies/TeacherInput" }
      responses:
        "200": { $ref: "#/components/responses/Teacher" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      tags: [teachers]
      operationId: deleteTeacher
      summary: Delete a teacher
      description: Fails with 409 `in_use` while curriculum items reference the teacher.
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/teachers/{id}/availability:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      tags: [teachers]
      operationId: getTeacherAvailability
      summary: Availability exceptions of a teacher
      responses:
        "200": { $ref: "#/components/responses/Availability" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [teachers]
      operationId: putTeacherAvailability
      summary: Replace the availability exceptions of a teacher
      requestBody: { $ref: "#/components/requestBodies/Availability" }
      responses:
        "200": { $ref: "#/components/responses/Availability" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }

  # ------------------------------------------------------------------ disciplines
...
  /api/time-grid:
    get:
      tags: [time-grid]
      operationId: getTimeGrid
      summary: The week shape
      responses:
        "200": { $ref: "#/components/responses/TimeGrid" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [time-grid]
      operationId: putTimeGrid
      summary: Set the week shape
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimeGrid" }
      responses:
        "200": { $ref: "#/components/responses/TimeGrid" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /api/periods:
    get:
      tags: [time-grid]
      operationId: listPeriods
      summary: List periods (pairs) with their times
      responses:
        "200":
          description: Periods ordered by number.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Period" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/periods/{number}:
    parameters:
      - name: number
        in: path
        required: true
        schema: { type: integer, minimum: 1, maximum: 16 }
    get:
      tags: [time-grid]
      operationId: getPeriod
      summary: Get a period
      responses:
        "200": { $ref: "#/components/responses/Period" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [time-grid]
      operationId: putPeriod
      summary: Create or replace a period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [starts_at, ends_at]
              properties:
                starts_at: { $ref: "#/components/schemas/Clock" }
                ends_at: { $ref: "#/components/schemas/Clock" }
      responses:
        "200": { $ref: "#/components/responses/Period" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      tags: [time-grid]
      operationId: deletePeriod
      summary: Delete a period
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------------ curriculum
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tt_access
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: tt_refresh
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
  schemas:
    ID: { type: integer, format: int64, minimum: 1 }
    Name: { type: string, minLength: 1, maxLength: 200 }
    Code:
      type: string
      minLength: 1
      maxLength: 32
      pattern: "^[a-z0-9][a-z0-9_-]*$"
    Clock:
      type: string
      pattern: "^([01][0-9]|2[0-3]):[0-5][0-9](:[0-5][0-9])?$"
      examples: ["08:30"]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [not_found] }
            message: { type: string }
...
    Room:
      type: object
      required: [id, building_id, name, room_type, capacity]
      properties:
        id: { $ref: "#/components/schemas/ID" }
        building_id: { $ref: "#/components/schemas/ID" }
        name: { $ref: "#/components/schemas/Name" }
        room_type: { $ref: "#/components/schemas/Code" }
        capacity: { type: integer, minimum: 1, maximum: 100000 }
    RoomInput:
      type: object
      additionalProperties: false
      required: [building_id, name, room_type, capacity]
      properties:
        building_id: { $ref: "#/components/schemas/ID" }
        name: { $ref: "#/components/schemas/Name" }
        room_type: { $ref: "#/components/schemas/Code" }
        capacity: { type: integer, minimum: 1, maximum: 100000 }
...
    Teacher:
      type: object
      required: [id, full_name, short_name]
      properties:
        id: { $ref: "#/components/schemas/ID" }
        full_name: { $ref: "#/components/schemas/Name" }
        short_name: { $ref: "#/components/schemas/Name" }
    TeacherInput:
      type: object
      additionalProperties: false
      required: [full_name, short_name]
      properties:
        full_name: { $ref: "#/components/schemas/Name" }
        short_name: { $ref: "#/components/schemas/Name" }
...
    Parity: { type: string, enum: [every, odd, even] }
    AvailabilityEntry:
      type: object
      additionalProperties: false
      required: [day, period, status]
      properties:
        day: { type: integer, minimum: 0, maximum: 6, description: 0 = Monday. }
        period: { type: integer, minimum: 1, maximum: 16 }
        parity: { allOf: [{ $ref: "#/components/schemas/Parity" }], default: every }
        status: { type: string, enum: [unavailable, undesired, preferred] }
    Availability:
      type: object
      additionalProperties: false
      required: [entries]
      description: Exceptions per slot; a slot without an entry is available with no preference.
      properties:
        entries:
          type: array
          maxItems: 336
          items: { $ref: "#/components/schemas/AvailabilityEntry" }
    TimeGrid:
      type: object
      additionalProperties: false
      required: [days, periods_per_day]
      properties:
        days: { type: integer, minimum: 1, maximum: 7 }
        periods_per_day: { type: integer, minimum: 1, maximum: 16 }
    Period:
      type: object
      required: [number, starts_at, ends_at]
      properties:
        number: { type: integer, minimum: 1, maximum: 16 }
        starts_at: { $ref: "#/components/schemas/Clock" }
        ends_at: { $ref: "#/components/schemas/Clock" }
...
  requestBodies:
...
    RoomInput:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RoomInput" }
...
    TeacherInput:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TeacherInput" }
...
    Availability:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Availability" }
...
  responses:
...
    Room:
      description: The room.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Room" }
...
    Teacher:
      description: The teacher.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Teacher" }
...
    Availability:
      description: The availability exceptions.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Availability" }
    TimeGrid:
      description: The time grid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TimeGrid" }
    Period:
      description: The period.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Period" }
...
    BadRequest:
      description: Malformed request (`invalid_request`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No valid session (`unauthorized`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The role may not do this (`forbidden`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The record does not exist (`not_found`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Unique key taken (`already_exists`), record still referenced (`in_use`) or schedule collision (`schedule_conflict`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unprocessable:
      description: Invalid field values (`validation_failed`), missing referenced record (`invalid_reference`) or database constraint (`constraint_violation`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...